created valdiateToken middleware and completed whole project

+++ controllers/contact.controller.js
@@ -2,16 +2,16 @@ const asyncHandler = require("express-async-handler");
 const Contact = require("../models/contact.model.js")
 //@desc get all contact
 //@route Get /api/contacts
-//@access public
+//@access private
 
 const getAllContacts = asyncHandler(async (req, res) => {
-  const contacts = await Contact.find()
+  const contacts = await Contact.find({user_id: req.user.id})
   res.status(200).json({contacts});
 });
 
 //@desc create contact
 //@route POST /api/contacts
-//@access public
+//@access private
 
 const createContact = asyncHandler(async (req, res) => {
   console.log("The request body is", req.body);
@@ -25,13 +25,14 @@ const createContact = asyncHandler(async (req, res) => {
     name,
     email,
     phone,
+    user_id: req.user.id
   })
   res.status(201).json(contact);
 });
 
 //@desc get contact
 //@route GET /api/contacts/:id
-//@access public
+//@access private
 
 const getContact = asyncHandler(async (req, res) => {
   
@@ -45,13 +46,17 @@ const getContact = asyncHandler(async (req, res) => {
 
 //@desc update contact
 //@route PUT /api/contacts/:id
-//@access public
+//@access private
 
 const updateContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if(!contact){
     res.status(404)
     throw new Error("Contact not found")
+  }
+  if(contact.user_id.toString()!= req.user.id){
+    req.status(403);
+    throw new Error("User don't have permission to update others contacts")
   }
   const updateContact = await Contact.findByIdAndUpdate(
     req.params.id,
@@ -63,7 +68,7 @@ const updateContact = asyncHandler(async (req, res) => {
 
 //@desc update contact
 //@route DELETE /api/contacts/:id
-//@access public
+//@access private
 
 const deleteContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
@@ -71,6 +76,11 @@ const deleteContact = asyncHandler(async (req, res) => {
     res.status(404)
     throw new Error("Contact not found")
   }
+  if(contact.user_id.toString()!= req.user.id){
+    req.status(403);
+    throw new Error("User don't have permission to update others contacts")
+  }
+  
   await Contact.findByIdAndDelete(req.params.id);
   res.status(200).json("contact deleted succesfully");
 });
